move Favorites functionality to FavoritesUtils
